+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - @since JSDoc tags to all library methods
+- method `addYears`
 
 ## [1.0.5] - 2019-07-17